requirements.txt
# QuantumShield — Production Dependencies
# =========================================

# ── Core Framework ──
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0

# ── Cryptography & TLS ──
sslyze>=5.2.0
cryptography>=42.0.0
cffi>=1.16.0

# ── Database ──
mysql-connector-python>=9.0.0

# ── Authentication & Security ──
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
flask-talisman>=1.1.0
Flask-Limiter>=3.5.0
argon2-cffi>=23.1.0

# ── Email ──
Flask-Mail>=0.9.1

# ── CBOM / Compliance Reporting ──
cyclonedx-python-lib>=7.0.0

# ── Environment ──
python-dotenv>=1.0.0

# ── Production WSGI ──
gunicorn>=21.2.0; sys_platform != 'win32'
waitress>=3.0.0; sys_platform == 'win32'

# ── Testing ──
pytest>=8.0.0
pytest-cov>=5.0.0

# ── Report Generation ──
weasyprint>=62.0; sys_platform != 'win32'
reportlab>=4.0

# ── Utilities ──
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.7.0
